move publish utility to tools dir
